refactor(mft): extract unified-parser tests to sibling module (<800 LOC)
WI-4.4 added the WTF-8 encoder, store_name_lossless, and their unit tests
to unified.rs, pushing it to 913 LOC — over the 800-LOC file-size gate.
Rather than rubber-stamp a file_size_exceptions.txt entry, extract the
test module to a sibling `unified/tests.rs` (mirroring the
compact_cache/tests.rs convention), bringing unified.rs back to 764 LOC.

The extracted `decode_name_u16_unpaired_surrogate_is_counted_and_replaced`
test asserted an exact delta on the process-global `lossy_name_count`
AtomicU64; that is order-dependent under parallel test execution (other
tests in the binary decode lossy names concurrently). Re-expressed as a
lower-bound (`>=`) assertion so it is concurrency-safe while still proving
the WI-4.1 tally rises on an unpaired surrogate.
